Rewritten file I/O classes to be thread-safe, cleaner and faster: CSFile (merged with CFile), CSFileText, CacheableScriptFile, CSVFile. Notably, CacheableScriptFile didn't really use the cached data when used with CResourceLock (at every lock, the script was reloaded). This is an important improvement, which grants now a faster script parsing and execution. Slightly reduced the startup time, thanks to the fixed script caching and because each script was actually read two times. Little optimizations of the LOS code, which is called very very often. Started to make CResourceLock thread-safe. Fixed 64 bits random number generation returning at maximum a 32 bits number.

+++ src/game/chars/CCharStatus.cpp
@@ -1329,8 +1329,8 @@ IT_TYPE CChar::CanTouchStatic( CPointMap &pt, ITEMID_TYPE id, CItem *pItem )
 	int x2 = pMapBlock->GetOffsetX(pt.m_x);
 	int y2 = pMapBlock->GetOffsetY(pt.m_y);
 
-	size_t iQty = pMapBlock->m_Statics.GetStaticQty();
-	for ( size_t i = 0; i < iQty; ++i )
+	uint iQty = pMapBlock->m_Statics.GetStaticQty();
+	for ( uint i = 0; i < iQty; ++i )
 	{
 		if ( !pMapBlock->m_Statics.IsStaticPoint(i, x2, y2) )
 			continue;
